manthan_v0_2_1/backend/requirements.txt
fastapi==0.110.*
uvicorn[standard]==0.30.*
pydantic<3
google-cloud-firestore>=2.16.0
openai>=1.40.0
